Single-core working

+++ include/hittable.h
@@ -2,13 +2,14 @@
 #define HITTABLE_H
 
 #include "common.h"
+#include "ray.h"
 
 typedef struct material Material;
 
 typedef struct hit_record {
   Vec3 p;
   Vec3 normal;
-  double t;
+  float t;
   bool front_face;
   Material *mat;
 } hit_record;
@@ -17,14 +18,16 @@ typedef struct hit_record {
 typedef bool (*HitFunction)(const void *object, const Ray *ray, Interval ray_t,
                             hit_record *record);
 
+// A hittable object struct containing a ptr to the object itself and the hit
+// function required
 typedef struct hittable {
-  void *object;             // Pointer to the actual object
-  HitFunction hit_function; // Function pointer for hit detection
+  void *object;              // Pointer to the actual object
+  HitFunction hit_function;  // Function pointer for specialized hit detection
 } Hittable;
 
+// Sets the hit record normal vector.
 void set_face_normal(hit_record *record, const Ray *r,
                      const Vec3 *outward_normal) {
-  // Sets the hit record normal vector.
   // NOTE: the parameter `outward_normal` is assumed to have unit length.
 
   record->front_face = vec3_dot(r->direction, *outward_normal) < 0;
