+++ include/interval.h
@@ -1,33 +1,32 @@
 #ifndef INTERVAL_H
 #define INTERVAL_H
-#include "math.h"
 #include <stdbool.h>
 
+#include "math.h"
+
 typedef struct interval {
-  double min;
-  double max;
+  float min;
+  float max;
 } Interval;
 
-Interval make_interval(double min, double max) {
+Interval make_interval(float min, float max) {
   Interval interval;
   interval.min = min;
   interval.max = max;
   return interval;
 }
 
-bool interval_contains(Interval *interval, double x) {
+bool interval_contains(Interval *interval, float x) {
   return interval->min <= x && x <= interval->max;
 }
 
-bool interval_surrounds(Interval *interval, double x) {
+bool interval_surrounds(Interval *interval, float x) {
   return interval->min <= x && x <= interval->max;
 }
 
-double interval_clamp(const Interval *interval, double x) {
-  if (x < interval->min)
-    return interval->min;
-  if (x > interval->max)
-    return interval->max;
+float interval_clamp(const Interval *interval, float x) {
+  if (x < interval->min) return interval->min;
+  if (x > interval->max) return interval->max;
   return x;
 }
